Struct matrix::format::banded::Banded

pub struct Banded<T: Element> {
    pub rows: usize,
    pub columns: usize,
    pub superdiagonals: usize,
    pub subdiagonals: usize,
    pub values: Vec<T>,
}

A banded matrix.

Fields
rows: usize

The number of rows.

columns: usize

The number of columns.

superdiagonals: usize

The number of superdiagonals.

subdiagonals: usize

The number of subdiagonals.

values: Vec<T>

The values of the diagonal elements stored as a (superdiagonals + 1 + subdiagonals) × columns matrix such that the first row corresponds to the uppermost superdiagonal whereas the last row corresponds to the lowest supdiagonal.

Methods

impl<T: Element> Banded<T>

fn new<S: Size>(size: S, superdiagonals: usize, subdiagonals: usize) -> Self

Create a zero matrix.

fn diagonals(&self) -> usize

Return the number of diagonals.

fn iter<'l>(&'l self) -> Iterator<'l, T>

Return a sparse iterator.