Struct matrix::format::compressed::Compressed

pub struct Compressed<T: Element> {
    pub rows: usize,
    pub columns: usize,
    pub nonzeros: usize,
    pub variant: Variant,
    pub values: Vec<T>,
    pub indices: Vec<usize>,
    pub offsets: Vec<usize>,
}

A compressed matrix.

Fields
rows: usize

The number of rows.

columns: usize

The number of columns.

nonzeros: usize

The number of nonzero elements.

variant: Variant

The format variant.

values: Vec<T>

The values of the nonzero elements.

indices: Vec<usize>

The indices of rows when variant = Column or columns when variant = Row of the nonzero elements.

offsets: Vec<usize>

The offsets of columns when variant = Column or rows when variant = Row such that the values and indices of the ith column when variant = Column or the ith row when variant = Row are stored starting from values[j] and indices[j], respectively, where j = offsets[i]. The vector has one additional element, which is always equal to nonzeros.

Methods

impl<T: Element> Compressed<T>

fn new<S: Size>(size: S, variant: Variant) -> Self

Create a zero matrix.

fn with_capacity<S: Size>(size: S, variant: Variant, capacity: usize) -> Self

Create a zero matrix with a specific capacity.

fn get<P: Position>(&self, position: P) -> T

Read an element.

fn set<P: Position>(&mut self, position: P, value: T)

Assign a value to an element.

Note that the function treats zero values as any other.

fn iter<'l>(&'l self) -> Iterator<'l, T>

Return a sparse iterator.

fn iter_mut<'l>(&'l mut self) -> IteratorMut<'l, T>

Return a sparse iterator allowing mutation.

fn resize<S: Size>(&mut self, size: S)

Resize the matrix.

fn retain<F>(&mut self, condition: F) where F: FnMut(usize, usize, &T) -> bool

Retain the elements that satisfy a condition and discard the rest.